+++ frontend/src/app/settings/example.txt
@@ -1,34 +1,5 @@
 
 
-<!-- Currency Section -->
-<div class="mt-5">
-  <div class="rounded-2xl border border-gray-200 bg-white p-5 dark:border-gray-800 dark:bg-white/[0.03] lg:p-6">
-    <h3 class="text-lg font-semibold text-gray-800 dark:text-white/90 lg:mb-1">
-      Currency Settings
-    </h3>
-    <div class="space-y-6">
-      <div class="p-5 border border-gray-200 rounded-2xl dark:border-gray-800 lg:p-6">
-        <div class="flex flex-col gap-5">
-          <div>
-            <label class="mb-2.5 block text-black dark:text-white">Select Currency</label>
-            <select
-              [ngModel]="selectedCurrency"
-              (change)="onCurrencyChange($event)"
-              class="w-full rounded-lg border border-gray-300 bg-transparent px-4 py-2.5 dark:border-gray-700">
-              <option *ngFor="let currency of currencies" [value]="currency.code">
-                {{currency.name}} ({{currency.symbol}})
-              </option>
-            </select>
-            <p class="mt-2 text-sm text-gray-500 dark:text-gray-400">
-              This will be the default currency for all your transactions
-            </p>
-          </div>
-        </div>
-      </div>
-    </div>
-  </div>
-</div>
-
 <!-- Payment Methods Section -->
 <div class="rounded-2xl border border-gray-200 bg-white p-5 mt-4 dark:border-gray-800 dark:bg-white/[0.03] lg:p-6">
   <div class="flex items-center justify-between mb-6">
