+++ frontend/src/app/settings/payment-methods/payment-methods.component.html
@@ -1,1 +1,90 @@
-<p>payment-methods works!</p>
+<!-- Payment Methods Section -->
+<div class="rounded-2xl border border-gray-200 bg-white p-5 mt-4 dark:border-gray-800 dark:bg-white/[0.03] lg:p-6">
+  <div class="flex items-center justify-between mb-6">
+    <div>
+      <h3 class="text-lg font-semibold text-gray-800 dark:text-white/90">Payment Methods</h3>
+      <p class="text-sm text-gray-500 dark:text-gray-400 mt-1">Manage your payment cards</p>
+    </div>
+    <button
+      (click)="addPaymentMethod()"
+      class="inline-flex items-center px-4 py-2 text-sm font-medium text-white bg-primary rounded-lg hover:bg-primary-dark transition-colors"
+    >
+      <svg class="w-4 h-4 mr-2" fill="none" viewBox="0 0 24 24" stroke="currentColor">
+        <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M12 6v6m0 0v6m0-6h6m-6 0H6" />
+      </svg>
+      Add Card
+    </button>
+  </div>
+
+  <!-- Payment Items Section -->
+  <div class="grid gap-4 md:grid-cols-2">
+    <div *ngFor="let method of paymentMethods"
+         class="relative group">
+      <div class="h-44 p-6 rounded-2xl bg-white border-2 border-gray-200 dark:bg-gray-800 dark:border-gray-700">
+        <div class="flex justify-between items-start">
+          <div>
+            <p class="text-gray-600 dark:text-gray-400 text-sm mb-4">
+              {{ method.type === 'visa' ? 'VISA' : 'MASTERCARD' }}
+            </p>
+            <p class="text-gray-900 dark:text-white font-medium tracking-widest text-lg">
+              •••• •••• •••• {{ getLastFourDigits(method.cardNumber) }}
+            </p>
+          </div>
+          <fa-icon
+            [icon]="getCardIcon(method.type)"
+            class="text-gray-600 dark:text-gray-400 text-2xl"
+          ></fa-icon>
+        </div>
+
+        <div class="absolute inset-0 rounded-2xl bg-gray-900/0 group-hover:bg-gray-900/5 dark:group-hover:bg-gray-900/20 transition-colors">
+          <div class="absolute right-4 top-4 flex -space-x-1 opacity-0 group-hover:opacity-100 transition-opacity">
+            <button (click)="setDefaultPaymentMethod(method.id)"
+                    *ngIf="method.isDefault"
+                    class="p-2 bg-white dark:bg-gray-700 rounded-full text-gray-600 dark:text-gray-400 hover:bg-gray-100 dark:hover:bg-gray-600 transition-colors">
+              <svg class="w-4 h-4" fill="none" viewBox="0 0 24 24" stroke="currentColor">
+                <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M5 13l4 4L19 7" />
+              </svg>
+            </button>
+            <button (click)="editPaymentMethod(method.id)"
+                    class="p-2 bg-white dark:bg-gray-700 rounded-full text-gray-600 dark:text-gray-400 hover:bg-gray-100 dark:hover:bg-gray-600 transition-colors">
+              <svg class="w-4 h-4" fill="none" viewBox="0 0 24 24" stroke="currentColor">
+                <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2"
+                      d="M15.232 5.232l3.536 3.536m-2.036-5.036a2.5 2.5 0 113.536 3.536L6.5 21.036H3v-3.572L16.732 3.732z" />
+              </svg>
+            </button>
+            <button (click)="deletePaymentMethod(method.id)"
+                    class="p-2 bg-white dark:bg-gray-700 rounded-full text-gray-600 dark:text-gray-400 hover:bg-gray-100 dark:hover:bg-gray-600 transition-colors">
+              <svg class="w-4 h-4" fill="none" viewBox="0 0 24 24" stroke="currentColor">
+                <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2"
+                      d="M19 7l-.867 12.142A2 2 0 0116.138 21H7.862a2 2 0 01-1.995-1.858L5 7m5 4v6m4-6v6m1-10V4a1 1 0 00-1-1h-4a1 1 0 00-1 1v3M4 7h16" />
+              </svg>
+            </button>
+          </div>
+        </div>
+
+        <div class="absolute bottom-6 left-6 right-6">
+          <div class="flex justify-between items-center">
+            <div>
+              <p class="text-gray-500 dark:text-gray-400 text-xs">Expires</p>
+              <p class="text-gray-700 dark:text-gray-300 text-sm">{{ method.expiryMonth }}/{{ method.expiryYear }}</p>
+            </div>
+            <span *ngIf="method.isDefault"
+                  class="inline-flex items-center rounded-full bg-gray-100 dark:bg-gray-700 px-2.5 py-0.5 text-xs font-medium text-gray-700 dark:text-gray-300">
+              Default
+            </span>
+          </div>
+        </div>
+      </div>
+    </div>
+  </div>
+
+  <!-- Empty State -->
+  <div *ngIf="paymentMethods.length === 0" class="text-center py-8">
+    <svg class="w-16 h-16 mx-auto text-gray-400" fill="none" viewBox="0 0 24 24" stroke="currentColor">
+      <path stroke-linecap="round" stroke-linejoin="round" stroke-width="1.5"
+            d="M3 10h18M7 15h1m4 0h1m-7 4h12a3 3 0 003-3V8a3 3 0 00-3-3H6a3 3 0 00-3 3v8a3 3 0 003 3z" />
+    </svg>
+    <h3 class="mt-4 text-sm font-medium text-gray-900 dark:text-white">No payment methods</h3>
+    <p class="mt-1 text-sm text-gray-500 dark:text-gray-400">Add a payment method to get started</p>
+  </div>
+</div>
